+++ llm_assistant/models/__init__.py
@@ -5,3 +5,4 @@ from . import llm_prompt
 from . import llm_prompt_tag
 from . import llm_prompt_category
 from . import llm_tool
+from . import llm_tool_bundle