+++ llm_assistant/tests/__init__.py
@@ -5,3 +5,4 @@ from . import test_llm_retry
 from . import test_stream_trace_sink
 from . import test_stream_duration_cap
 from . import test_effective_tools
+from . import test_tool_resolver